% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaCI.R
\name{bcaCI}
\alias{bcaCI}
\title{Bootstrap Confidence Interval for Model Comparison}
\usage{
bcaCI(
  data,
  predicttime,
  formula1,
  formula2,
  model,
  cause,
  method,
  nboot,
  alpha
)
}
\arguments{
\item{data}{observed data. should include time, cause indicator with 0 for censoring, and all covariates to be included in the prognostic models.}

\item{predicttime}{time of model evaluation}

\item{formula1}{list of Hist formulas for cause-1 and cause-2 events in the first model. cause-1 and cause-2 can have different formulas. If only one formula is given, will be used for both cause-1 and cause-2.}

\item{formula2}{list of Hist formulas for cause-1 and cause-2 events in the second model. cause-1 and cause-2 can have different formulas. If only one formula is given, will be used for both cause-1 and cause-2.}

\item{model}{vector of two prognostic models to be fitted for comparison. "cox" and "fg" can be fitted. If only one model is specified, will be used for both models for comparison.}

\item{cause}{1 or 2 for cause-specific evaluation. 0 for overall assessment.}

\item{method}{either "pdi" or "ExC"}

\item{nboot}{number of bootstrap samples to be taken.}

\item{alpha}{specify confidence level.}
}
\value{
bias-corrected and accelerated confidence interval of difference between two models at specified confidence level.
}
\description{
Confidence interval of difference between two models based on bias-corrected and accelerated bootstrap method
}
