% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExC.R
\name{ExC}
\alias{ExC}
\title{Extended Concordance Index for Overall Evaluation}
\usage{
ExC(time, delta, cif1, cif2, surv, predicttime)
}
\arguments{
\item{time}{The observed time or followed up time for right censored data.}

\item{delta}{The observed status indicator. 0 for censoring.}

\item{cif1}{Cause-1 CIF function estimated at predict time.}

\item{cif2}{Cause-2 CIF function estimated at predict time.}

\item{surv}{Survival function estimated at predict time.}

\item{predicttime}{A single time point of prediction or evaluation.}
}
\value{
Estimated concordance index for overall evaluation at predicttime.
}
\description{
Estimate extended concordance index that evaluate the ability of prognostic models to discriminate subjects from all groups simultaneously and to predict event status under competing risks setting.
}
