% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExCl.R
\name{ExCl}
\alias{ExCl}
\title{Extended Concordance Index for Cause-specific Evaluation}
\usage{
ExCl(time, delta, cif1, cif2, surv, predicttime, cause)
}
\arguments{
\item{time}{The observed time or followed up time for right censored data.}

\item{delta}{The observed status indicator. 0 for censoring.}

\item{cif1}{Cause-1 CIF function estimated at predict time.}

\item{cif2}{Cause-2 CIF function estimated at predict time.}

\item{surv}{Survival function estimated at predict time.}

\item{predicttime}{A single time point of prediction or evaluation.}

\item{cause}{Specify cause of interest.}
}
\value{
Estimated concordance index for cause-specific evaluation at predicttime.
}
\description{
Estimate extended concordance index that evaluate the ability of a prognostic model to distinguish between the cause of interest and the healthy control and to predict event status under competing risks setting.
}
